Conjunction Integration Tests
- Consolidate for sake of Spicerack spec includes
