--- conjunction/spec/support/test_classes/conjunctions.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module GalacticFederation
-  class EarthlingFleeb < ApplicationFleeb; end
-  class SatelliteFleeb < ApplicationFleeb; end
-end
-
-class PlanetaryAddressFleeb < ApplicationFleeb; end
-
-class StarlightFleeb < ApplicationFleeb; end
-class GenericFleeb < ApplicationFleeb; end
